refactor: Refatorar testes de confirmação
refatorar testes de aprovação e recusa de credito de rubis
